backend/requirements.txt
anyio==4.6.0
async-timeout==4.0.3
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.92.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
loguru==0.6.0
passlib==1.7.4
pip==24.2
psycopg2-binary
pydantic==1.10.2
PyJWT==2.9.0
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.48
starlette==0.25.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.20.0
wheel==0.44.0
